Move uri resolvers+wiring to uri
