+++ grpc-api/src/main/java/io/deephaven/grpc_api/runner/DeephavenApiServer.java
@@ -10,11 +10,11 @@ import io.deephaven.grpc_api.appmode.ApplicationInjector;
 import io.deephaven.grpc_api.console.ConsoleServiceGrpcImpl;
 import io.deephaven.grpc_api.log.LogInit;
 import io.deephaven.grpc_api.session.SessionService;
-import io.deephaven.grpc_api.uri.UriResolver;
-import io.deephaven.grpc_api.uri.UriResolvers;
 import io.deephaven.internal.log.LoggerFactory;
 import io.deephaven.io.logger.Logger;
-import io.deephaven.grpc_api.uri.UriResolversInstance;
+import io.deephaven.uri.resolver.UriResolver;
+import io.deephaven.uri.resolver.UriResolvers;
+import io.deephaven.uri.resolver.UriResolversInstance;
 import io.deephaven.util.annotations.VisibleForTesting;
 import io.deephaven.util.process.ProcessEnvironment;
 import io.deephaven.util.process.ShutdownManager;
