+++ grpc-api/src/main/java/io/deephaven/grpc_api/uri/QueryScopeResolver.java
@@ -3,6 +3,8 @@ package io.deephaven.grpc_api.uri;
 import io.deephaven.grpc_api.console.GlobalSessionProvider;
 import io.deephaven.uri.DeephavenUri;
 import io.deephaven.uri.QueryScopeUri;
+import io.deephaven.uri.resolver.UriResolver;
+import io.deephaven.uri.resolver.UriResolversInstance;
 
 import javax.inject.Inject;
 import java.net.URI;
