+++ java-client/uri/src/main/java/io/deephaven/uri/resolver/UriResolversInstance.java
@@ -1,4 +1,4 @@
-package io.deephaven.grpc_api.uri;
+package io.deephaven.uri.resolver;
 
 import java.util.Objects;
